Preliminary Hookup
The CodeEditor component has a basic link and text_model, in order to communicate with the emulation core and parser(soon to be added).

Execute Button is just a button, but the goal is to create a Datapath that won't get reset upon another but instead would execute the next instruction.