Lock rust version (#105)
This locks the rust version to the most recent available (1.67.0). Future releases will be ignored by GitHub Actions until this version is manually updated. The rust-toolchain.toml file also forces clients to have the specific version of the toolchain for the project to avoid compatibility issues.

In addition, the GitHub Action for installing the rust toolchain has been replaced with `rustup show`. The action is no longer actively developed as of the time of commit and does not fully support reading from the rust-toolchain.toml file from testing. While the functionality of `rustup show` installing from the toolchain file is not documented well, it will at the moment perform the same action. Should a proper "install the toolchain from the rust-toolchain.toml file" feature be released, this would be preferred.

Finally, as a bonus, *all* tests in check.yml will now run regardless of how each test completed. This, for example, allows for testing the build of the project even if there was a small formatting issue reported by `cargo fmt`. Previously, a formatting issue would have stopped all checking and possibly hide other problems that could be fixed at the same time.