Implementing Load File
Text Model is not given the interior mutability treatment in order to borrow it for loading in a file.

Restructured the placements for some variables (i.e. console component) and added more comments and made clarifications.

Introduced futures to prevent certain events from happening out of order, resulting in race conditions.