87 support beq and bne core (#109)
* Added initial stuff to implement the J instruction

* datapath: Add reset()

* Console and regview branch (#104)

* made some boxes

basic stuff

* aside

* add float to move it right

* making table

* dynamic size

* using better units

* Create tabs.rs

* screen fitting

* resolve merge conflict

* resolve merge conflict 2

* progress on regview

* changed according to clippy

* update

* reformatted - clippy

* Lock rust version (#105)

This locks the rust version to the most recent available (1.67.0). Future releases will be ignored by GitHub Actions until this version is manually updated. The rust-toolchain.toml file also forces clients to have the specific version of the toolchain for the project to avoid compatibility issues.

In addition, the GitHub Action for installing the rust toolchain has been replaced with `rustup show`. The action is no longer actively developed as of the time of commit and does not fully support reading from the rust-toolchain.toml file from testing. While the functionality of `rustup show` installing from the toolchain file is not documented well, it will at the moment perform the same action. Should a proper "install the toolchain from the rust-toolchain.toml file" feature be released, this would be preferred.

Finally, as a bonus, *all* tests in check.yml will now run regardless of how each test completed. This, for example, allows for testing the build of the project even if there was a small formatting issue reported by `cargo fmt`. Previously, a formatting issue would have stopped all checking and possibly hide other problems that could be fixed at the same time.

* Add more to pipline for J support

* Completed calculation of jump address

* Implemented the J instruction

Need to write test cases

* Actually implemented the J instruction, for real

* Wrote basic test case for jump instruction

* Done J instruction tests

* Increased memory size to 64k

* Added AluZ datapath signal

* Added branch decision stuff to datapath

* Updated comment

* Implemented beq and bne

* Fixed bug with branch addr calc

* Added branch test cases, one fails

* datapath: Fix BEQ bug

* Ran cargo fmt

---------

Co-authored-by: Jerrett Longworth <jerrett7@gmail.com>
Co-authored-by: huynguyen2019 <60188307+huynguyen2019@users.noreply.github.com>