UI branch: Load File Implemented (#106)
* Preliminary Hookup

The CodeEditor component has a basic link and text_model, in order to communicate with the emulation core and parser(soon to be added).

Execute Button is just a button, but the goal is to create a Datapath that won't get reset upon another but instead would execute the next instruction.

* Rewrite

Still have ways to go, but here we do not have to rely too much on crates and wasting more time on research.

* Assemble and Execute

Has the parser and emulation core linked together with basic output to the console.

* cleanup

* merge with main

Should be ready for review...

* cargo fmt

* Complied with Clippy

* ...

* Fixed structure of components

* Implementing Load File

Text Model is not given the interior mutability treatment in order to borrow it for loading in a file.

Restructured the placements for some variables (i.e. console component) and added more comments and made clarifications.

Introduced futures to prevent certain events from happening out of order, resulting in race conditions.

* fmt and clippy

* Committed The Ultimate Yew Sin

In order to actually show what is going on inside the Datapath, we have to force update it. This will apply to anything that involves with the Datapath (and maybe the parser).

* cargo fmt

---------

Co-authored-by: Kevin Cahalan <69361467+kevinacahalan@users.noreply.github.com>